Начала прибираться в компонентах

+++ Lambda/Core/GreedyComponent/GreedyComponentPunter.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using Core.Components;
 using Core.Objects;
 
 namespace Core.GreedyComponent
@@ -139,7 +140,7 @@ namespace Core.GreedyComponent
             var nodeToComponent = BuildNodeToComponent();
 
             desiredComponent = new HashSet<Component> {first};
-            var S = new SortedSet<(int, Component)>();
+            var S = new SortedSet<(long, Component)>();
             foreach (var c in first.Nodes.SelectMany(n => map.GetAvaliableEdges(n.Id, punter).Select(e => nodeToComponent[e.Item1.Id])).Distinct())
             {
                 S.Add((first.Mines.Union(c.Mines).Sum(x => c.Scores[x.Id] + desiredScores[x.Id]), c));
